refactor(ptt): extract client port runtimes
Why:

- ClientApp was still carrying the LLM HTTP adapter and Injection worker runtime after the port abstraction landed, which kept app.rs responsible for adapter details it only coordinates.

- Moving those runtimes behind llm and injector_runtime modules makes ClientPorts the boundary and keeps main.rs focused on CLI parsing plus adapter construction.

Side effects:

- Adds direct LLM in-memory stream tests and an injector subprocess timeout regression test in the extracted modules.
